+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   CollectionAdapter for (#1509)
 - Introduced new generator parameter `scala3_sources` (also available via
   scalapb.proto) to generate sources that are compatible with Scala 3 (#1576)
+- aux_*_options now accept "*" in the target field to match all entities in
+  scope.
 
 ## [0.11.13]
 - Added input and output message type for method descriptor (#1503)